requirements.txt
Pillow==10.4.0
pynput==1.7.7
Requests==2.32.3
python-dotenv